1.move all java config files to the 'config' folder. 2.change the test classes related to the config classes 3. add IPerformance interface and change the Performance class into an interface-based implementation.
