+++ src/main/java/com/springinaction/concert/IPerformance.java
@@ -1,0 +1,9 @@
+package com.springinaction.concert;
+
+/**
+ * Created by Johnny on 11/23/2016.
+ */
+public interface IPerformance {
+    void perform();
+    void anotherPerform();
+}
