+++ src/main/java/com/springinaction/concert/Performance.java
@@ -3,7 +3,7 @@ package com.springinaction.concert;
 /**
  * Created by Johnny on 11/22/2016.
  */
-public class Performance {
+public class Performance implements IPerformance{
     public void perform() {
         System.out.println("The play is performing.");
     }
